Module #3. FINAL MODULE TASK. Attempt #2.

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -3,7 +3,7 @@ package lang.print.gaps.finalModuleTask;
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
 
-        int temperatureFahrenheit = (temperatureCelsius * 9 / 5) + 32;
+        float temperatureFahrenheit = (temperatureCelsius * 9 / 5f) + 32;
         System.out.println(temperatureFahrenheit);
     }
 }
